Add authentication for comment

+++ api/service/comment.py
@@ -2,16 +2,19 @@ from flask import Blueprint, request, abort, jsonify
 from service.db import DB
 from utils import sha256
 from datetime import datetime
+from service.auth import token_required
 
 db = DB()
 
 comment_bp = Blueprint("comment", __name__)
 
 
+def getComment(commentID):
+    return db.search('Comment', {'comment_id': commentID})
+
 # get all comments
 @comment_bp.route("/comments")
 def get_comments():
-
     return db.search('Comment')
 
 # search for certain comments
@@ -27,10 +30,11 @@ def search_comment():
         return db.search('Comment', request.args.copy())
     except:
         return "Incorrect input"
-  
+
 # add a comment
 @comment_bp.route("/comment", methods=['POST'])
-def add_comments():
+@token_required
+def add_comments(current_user):
 
     colNames = db.getColumnNames("Comment")
 
@@ -38,7 +42,8 @@ def add_comments():
         body = request.json.copy()
     except:
         body = request.form.copy()
-    
+
+    body['user_name'] = current_user['user_name']
     for colName in colNames:
         if body.get(colName) is None and colName != "comment_id":
             return "Missing " + colName
@@ -48,23 +53,36 @@ def add_comments():
         for value in body.values():
             s += value
         body["comment_id"] = sha256(s + str(datetime.now()))
-    
+
     return db.insert("Comment", body)
 
 # delete comment
 @comment_bp.route("/comment/<comment_id>", methods=['DELETE'])
-def delete_comments(comment_id):
-
+@token_required
+def delete_comments(current_user, comment_id):
+    comment = getComment(comment_id)
+    if not comment:
+        return "No comment with comment id " + comment_id + " found"
+    if comment['user_name'] != current_user['user_name']:
+        return "Not authorized to delete this comment"
+    
     db.delete("Comment", {"comment_id": comment_id})
 
     return "Comment id " + comment_id + " deleted"
 
 # update comment
 @comment_bp.route("/comment/<comment_id>", methods=['PUT'])
-def update_comments(comment_id):
+@token_required
+def update_comments(current_user, comment_id):
+    comment = getComment(comment_id)
+    if not comment:
+        return "No comment with comment id " + comment_id + " found"
+    if comment['user_name'] != current_user['user_name']:
+        return "Not authorized to update this comment"
+
     try:
         body = request.json.copy()
     except:
         body = request.form.copy()
-    
+
     return db.update("Comment", {"comment_id": comment_id}, body)
